requirements.txt
django
djangorestframework
reportlab
psycopg2-binary